Testing And Debugging

+++ controllers/author.js
@@ -1,20 +1,33 @@
-
 const Author = require('../models/author');
+const logger = require('../utils/logger');
+
+const nameRegex = /^[a-zA-Z0-9]+$/;
 
 const createAuthor = async (req, res) => {
-    const { name } = req.body;
-    if (!name || name === '') {
-        return res.status(400).json({ data: {}, message: 'Name is required', success: false });
+    const { firstName, lastName, authorID } = req.body;
+    const { username, isAdmin } = req.user;
+
+    if (!isAdmin) {
+        logger.warn(`Unauthorized attempt to create author by ${username}`);
+        return res.status(403).json({ data: {}, message: 'Unauthorized', success: false });
+    }
+
+    if (!firstName || firstName === '' || !lastName || lastName === '') {
+        return res.status(400).json({ data: {}, message: 'All fields are required', success: false });
+    }
+    if (!nameRegex.test(firstName) || !nameRegex.test(lastName)) {
+        return res.status(400).json({ data: {}, message: 'Invalid name', success: false });
     }
     try {
-        const existingAuthor = await Author.findOne({ where: { name } });
+        const existingAuthor = await Author.findOne({ where: { firstName, lastName } });
         if (existingAuthor) {
             return res.status(409).json({ data: {}, message: 'Author already exists', success: false });
         }
-        const newAuthor = await Author.create({ name, authorID: 'AU' + Math.floor(Math.random() * 1000) });
+        const newAuthor = await Author.create({ firstName, lastName, authorID });
         return res.status(201).json({ data: newAuthor, message: 'Author created', success: true });
-    } catch (error) {
-
+    }
+    catch (error) {
+        logger.error(error.message);
         return res.status(500).json({ data: {}, message: error.message, success: false });
     }
 };
@@ -27,6 +40,7 @@ const getAllAuthors = async (req, res) => {
         }
         return res.status(200).json({ data: authors, message: 'Authors retrieved', success: true });
     } catch (error) {
+        logger.error(error.message);
         return res.status(500).json({ data: {}, message: error.message, success: false });
     }
 };
@@ -42,42 +56,59 @@ const getAuthorById = async (req, res) => {
         }
         return res.status(404).json({ data: {}, message: 'Author with the specified ID does not exist', success: false });
     } catch (error) {
+        logger.error(error.message);
         return res.status(500).json({ data: {}, message: error.message, success: false });
     }
 }
 
 const updateAuthor = async (req, res) => {
     const { id } = req.params;
-    const { name } = req.body;
+    const { firstName, lastName } = req.body;
 
-    if (!name || name === '') {
-        return res.status(400).json({ data: {}, message: 'Name is required', success: false });
+    if (!firstName || firstName === '' || !lastName || lastName === '') {
+        return res.status(400).json({ data: {}, message: 'All fields are required', success: false });
     }
+    if (!nameRegex.test(firstName) || !nameRegex.test(lastName)) {
+        return res.status(400).json({ data: {}, message: 'Invalid name', success: false });
+    }
+
     try {
-        const [updated] = await Author.update(req.body, {
+        const author = await Author.findOne({ where: { authorID: id } });
+        if (!author) {
+            return res.status(404).json({ data: {}, message: 'Author not found', success: false });
+        }
+        const updatedAuthor = await Author.update({ firstName, lastName }, {
             where: { authorID: id }
         });
-        if (updated) {
-            const updatedAuthor = await Author.findOne({ where: { authorID: id } });
-            return res.status(200).json({ data: updatedAuthor, message: 'Author updated', success: true });
-        }
-        throw new Error('Author not found');
+        return res.status(200).json({ data: updatedAuthor, message: 'Author updated', success: true });
     } catch (error) {
+        logger.error(error.message);
         return res.status(500).json({ data: {}, message: error.message, success: false });
     }
 };
 
 const deleteAuthor = async (req, res) => {
     const { id } = req.params;
+    const { username, isAdmin } = req.user;
+
+    if (!isAdmin) {
+        logger.warn(`Unauthorized attempt to delete author by ${username}`)
+        return res.status(403).json({ data: {}, message: 'Unauthorized', success: false });
+    }
     try {
+        if (!await Author.findOne({ where: { authorID: id } })) {
+            return res.status(404).json({ data: {}, message: 'Author not found', success: false });
+        }
+
         const deleted = await Author.destroy({
             where: { authorID: id }
         });
         if (deleted) {
+            logger.info(`Author with ID ${id} deleted by ${username}`);
             return res.status(202).json({ data: {}, message: 'Author deleted', success: true });
         }
-        throw new Error("Author not found");
     } catch (error) {
+        logger.error(error.message);
         return res.status(500).json({ data: {}, message: error.message, success: false });
     }
 };
